requirements.txt
# Core Deep Learning
torch>=2.0.0

# Data Processing & Numerical Computing
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning & Metrics
scikit-learn>=1.0.0

# Utilities
tqdm>=4.62.0
PyYAML>=5.4.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.12.0
